+++ metadata_backend/server.py
@@ -10,7 +10,10 @@ import uvloop
 from aiohttp import web
 from cryptography.fernet import Fernet
 
+from metadata_backend.message_broker.mq_service import MQPublisher
+
 from .api.auth import AAIServiceHandler, AccessHandler
+from .api.handlers.files import FilesAPIHandler
 from .api.handlers.object import ObjectAPIHandler
 from .api.handlers.publish import PublishSubmissionAPIHandler
 from .api.handlers.rems_proxy import RemsAPIHandler
@@ -68,6 +71,7 @@ async def init(
     datacite_handler = DataciteServiceHandler()
     rems_handler = RemsServiceHandler()
     aai_handler = AAIServiceHandler()
+    mq_publisher = MQPublisher()
 
     async def close_http_clients(_: web.Application) -> None:
         """Close http client session."""
@@ -91,13 +95,17 @@ async def init(
         metax_handler=metax_handler, datacite_handler=datacite_handler, rems_handler=rems_handler
     )
     _publish_submission = PublishSubmissionAPIHandler(
-        metax_handler=metax_handler, datacite_handler=datacite_handler, rems_handler=rems_handler
+        metax_handler=metax_handler,
+        datacite_handler=datacite_handler,
+        rems_handler=rems_handler,
+        mq_publisher=mq_publisher,
     )
     _user = UserAPIHandler()
     _xml_submission = XMLSubmissionAPIHandler(
         metax_handler=metax_handler, datacite_handler=datacite_handler, rems_handler=rems_handler
     )
     _template = TemplatesAPIHandler()
+    _file = FilesAPIHandler()
     api_routes = [
         # retrieve workflows
         web.get("/workflows", _common_api_handler.get_workflows),
@@ -128,6 +136,8 @@ async def init(
         web.get("/submissions", _submission.get_submissions),
         web.post("/submissions", _submission.post_submission),
         web.get("/submissions/{submissionId}", _submission.get_submission),
+        web.get("/submissions/{submissionId}/files", _submission.get_submission_files),
+        web.post("/submissions/{submissionId}/files", _submission.add_submission_files),
         web.put("/submissions/{submissionId}/doi", _submission.put_submission_path),
         web.put("/submissions/{submissionId}/dac", _submission.put_submission_path),
         web.patch("/submissions/{submissionId}", _submission.patch_submission),
@@ -141,6 +151,8 @@ async def init(
         web.post("/submit/{workflow}", _xml_submission.submit),
         # validate
         web.post("/validate", _xml_submission.validate),
+        # File operations
+        web.get("/files", _file.get_project_files),
     ]
     _rems = RemsAPIHandler(metax_handler=metax_handler, datacite_handler=datacite_handler, rems_handler=rems_handler)
     api_routes.append(web.get("/rems", _rems.get_workflows_licenses_from_rems))